fix hash code for LongIntHashMap
Before it was computing the murmur3 hash and then ignoring
the result and using `Long.hashCode`. This change moves
the computation of the murmur3 hash to a helper function
and now LongIntHashMap uses that value.
